Add option for single / normal view of result
By default, the wiki search result link opens the tiddler in "single
page view", so nothing is loaded, but only the tiddler. It's fast and
usefull, but sometimes you may want to open it in "normal view" (to edit
for example), thus I added a link to this (click on the `#` icon).